fixed a bug(payload tool - crash upon deleting non selected item)

+++ SwitchHacksAllInOne/PayloadTools/MainTool.xaml.cs
@@ -134,11 +134,14 @@ namespace SHAIO.PayloadTools
         private void RemoveButton(object sender, RoutedEventArgs e)
         {
             FileInfo fs = (FileInfo)Combo.SelectedItem;
-            MessageBoxResult result = MessageBox.Show("Do you want to delete " + fs.Name + "?", "Delete", MessageBoxButton.YesNo);
-            if (result == MessageBoxResult.Yes)
+            if (fs != null)
             {
-                File.Delete(fs.FullName);
-                FillItemSource();
+                MessageBoxResult result = MessageBox.Show("Do you want to delete " + fs.Name + "?", "Delete", MessageBoxButton.YesNo);
+                if (result == MessageBoxResult.Yes)
+                {
+                    File.Delete(fs.FullName);
+                    FillItemSource();
+                }
             }
         }
     }
